+++ src/mocks/Mocks.js
@@ -10,35 +10,26 @@ Unless required by applicable law or agreed to in writing, software distributed 
 */
 
 const Behaviors = require("./Behaviors");
-const FilesHandler = require("./FilesHandler");
 const FixturesHandler = require("./FixturesHandler");
 const FixtureHandler = require("./FixtureHandler");
 const Fixtures = require("./Fixtures");
 
 class Mocks {
-  constructor(settings, eventEmitter, core) {
-    this._settings = settings;
+  constructor(eventEmitter, settings, loaders, core) {
     this._eventEmitter = eventEmitter;
+    this._settings = settings;
+    this._loaders = loaders;
+
     this._fixturesHandler = new FixturesHandler(core);
     this._fixturesHandler.addHandler(FixtureHandler);
-    this._filesHandler = new FilesHandler(this._settings, this._eventEmitter);
-    this._fixtures = new Fixtures(this._filesHandler, this._settings, this._eventEmitter);
-    this._behaviors = new Behaviors(this._filesHandler, this._settings, this._eventEmitter);
+    this._fixtures = new Fixtures(this._loaders, this._eventEmitter);
+    this._behaviors = new Behaviors(this._loaders, this._settings, this._eventEmitter);
   }
 
   async init() {
-    await this._filesHandler.init();
     await this._fixtures.init(this._fixturesHandler);
     await this._behaviors.init(this._fixturesHandler, this._fixtures);
     return Promise.resolve();
-  }
-
-  async stop() {
-    await this._filesHandler.stop();
-  }
-
-  async start() {
-    await this._filesHandler.start();
   }
 
   addFixturesHandler(Handler) {